Allow inheriting decorated class
If an operator class was decorated and the operator class is subclassed
by other class, it causes errors because current decorator returns
a function, not class.
   TypeError: function() argument 1 must be code, not str

Example>
  @input("image", Image, IOType.IN_MEMORY)
  @output("seg_image", Image, IOType.IN_MEMORY)
  class MonaiInferenceOperator(Operator):
      ...

  class SpleenSegOperator(MonaiInferenceOperator):
      ...

This patch solves the issue by returning class instead of a method.
It also use a special builder method `_builder()` to chain methods
to set an operator or an application.