Fix inheritance issue for Operator and Application
Previously, a new type that overrides was returned by @input @output
@env decorators which can cause the following error when executing
super() method in __init__():

TypeError: super(type, obj): obj must be an instance or subtype of type

We don't need to create a new type (with the same name).
Instead, just override _builder property to achieve the same behavior.

Also, this patch does not set _env in the constructor of Application
and Operator. Instead it is set to None by default in the class
definition and check if has a value or not in @env decorator.