Fix DICOM attribute handling
Existing implementation did not set attributes in
advance, and their availability are checked by
handing AttributeError.

For the reason, initializing those attributes with
None caused spleen app execution failure.
(All metadata dictionary in Domain class is filled
with None even for non-available attributes)

With this patch, the attribute's availability is checked
by checking if its value is None and getters return
None if attributes are not set (using getattr() method).

Signed-off-by: Gigon Bae <gbae@nvidia.com>