CI: add build and push models workflow
This workflow will build and push into the ghcr.io the
example MAP in examples/apps/simple_imaging_app.

This workflow will use terraform to launch a VM with a GPU and then run
the monai-deploy-sdk package subcommand inside it.

NOTE: you need to setup the secrets:
* AZURE_CLIENT_ID
* AZURE_SUBSCRIPTION_ID
* AZURE_TENANT_ID
* AZURE_CLIENT_SECRET

The workflow will patch holoscan, so that it works when there is no
cache.

The workflow will install a "patched" libseccomp package so that we can
install libnvidia-container.

The workflow will use nvidia docker runtime to build, load, export, ...
the image.

The workflow uses the smallest/cheapest Azure Image with GPU available
in Western Europe, which is Standard_NC4as_T4_v3. For this to work,
you need to request a quota increase to Azure Help Desk.

However, the default 30GB disc is not enough for the build, so you need
to setup a 64GB.

:WARNING: This will incur in costs in Azure Cloud, use it with caution.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.com>