Address concerns that the rescaled pixel array from DICOM is of type float64.
In the last version, rescale slope and intercept values were tried to cast to
the smallest can_cast Nunpy type. Using pydicom apply_rescale loses this as it
returns float64 if rescale slope and intercept are applied. So, enhancing the
code will test for equivalence before "unsafely" casting the data. This will
NOT reduce the mem usage of this operator as a new array will be created by
casting from float64, and there is a small overhead for the checks. However,
there are cases where downstream custom operators may save memory footprint.

Signed-off-by: M Q <mingmelvinq@nvidia.com>